requirements.txt
tensorflow
numpy
matplotlib
streamlit
Pillow